add experiences, optimize images size
