+++ src/data/experiences.js
@@ -1,0 +1,23 @@
+export const experiences = [
+  {
+    id: 1,
+    company: "Appier",
+    title: "Frontend Engineer Intern",
+    startDate: "Jun. 2024",
+    endDate: "Aug. 2024",
+    responsibilities: ["To be updated"],
+    image: "/appier.svg"
+  },
+  {
+    id: 2,
+    company: "GBCSGroup",
+    title: "Backend Developer Intern",
+    startDate: "Feb. 2024",
+    endDate: "May 2024",
+    responsibilities: ["Led the backend development of GHG project, a platform to help companies calculate their greenhouse gas emissions",
+                  "Developed a RESTful API with Python FastAPI framework",
+                  "Actively engaged in software development life cycle, including project planning, customer requirement analysis, API design, documentation, QA, and maintenance",
+                  "Collaborated with cross-functional teams to identify strengths and weaknesses of GHG models through analysis of more than 40 equations and 400 endpoints across fleet modes and lifecycles"],
+    image: "gbcs_group.jpeg"
+  }
+]
